Use result after validation
In some cases, passing additional options to the validator causes the
incoming data to be changed.

For example, adding the enjoi settings:

    allowUnknown: true
    stripUnknown: true

Will cause any incoming, unknown properties to be removed from the
object that is being validated.

This change ensures those results are used rather than the original,
pre-validated object.